adding booth view
