+++ src/component/OthersComponents/WorkInProgress.jsx
@@ -1,0 +1,9 @@
+function WorkInProgress() {
+  return (
+    <div className="box" id="not-results-box">
+      <p className="is-size-3 has-text-weight-bold">¡Trabajo en progreso!</p>
+    </div>
+  );
+}
+
+export default WorkInProgress;
