+++ src/pages/Booth/Panel/LoggerBooth.jsx
@@ -1,0 +1,7 @@
+import WorkInProgress from "../../../component/OthersComponents/WorkInProgress";
+
+function LoggerBoth() {
+  return <WorkInProgress />;
+}
+
+export default LoggerBoth;
