+++ src/pages/Booth/Panel/StatisticsBooth.jsx
@@ -1,0 +1,7 @@
+import WorkInProgress from "../../../component/OthersComponents/WorkInProgress";
+
+function StatisticsBooth() {
+  return <WorkInProgress />;
+}
+
+export default StatisticsBooth;
