+++ src/pages/Booth/Panel/VerifyElection.jsx
@@ -1,0 +1,12 @@
+import { useEffect } from "react";
+import { useState } from "react";
+import { useParams } from "react-router-dom";
+import FooterParticipa from "../../../component/Footers/FooterParticipa";
+import TitlePsifos from "../../../component/OthersComponents/TitlePsifos";
+import WorkInProgress from "../../../component/OthersComponents/WorkInProgress";
+
+function VerifyElection() {
+  return <WorkInProgress />;
+}
+
+export default VerifyElection;
